engine/requirements.txt
torch
transformers
faster-whisper
pyaudio
numpy
openai
python-dotenv
elevenlabs
pygame
